+++ core/server/index.ts
@@ -24,3 +24,4 @@ export * from './usecases/search.js';
 export * from './usecases/status.js';
 export * from './usecases/tags.js';
 export * from './usecases/thumbnail.js';
+export * from './usecases/variants.js';